chore: changed float to BigDecimal on dailyPrice

+++ src/main/java/br/com/fiap/Shift/models/Vehicle.java
@@ -6,6 +6,8 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Entity
 public class Vehicle {
@@ -15,5 +17,5 @@ public class Vehicle {
     private String model;
     private Integer year;
     private String licensePlate;
-    private float dailyPrice;
+    private BigDecimal dailyPrice;
 }
